Initium version Inanis Whitepaper (Initium Inanis)
This Whitepaper describes the main ideas behind the Initium Inanis (version Inanis). Inanis Version of Initium is the test version of the Initium blockchain which is to be considered the lab version. The codes of this version are available on Initium Foundations GitHub. The next version is the Initium Primus and its release date will be announced by the Initium Foundation. The Primus version will enable the test net of the Initium blockchain. The successful tests of Initium Primus lead us to release the main version, the Initium Genesis.