Sets the status between validators and ancestor nodes
Number of validators to sample for the ancestor repair pub const ANCESTOR_HASH_REPAIR_SAMPLE_SIZE: usize = 21;
Even assuming 20% of validators are malicious, the chance that >= 11 of the ANCESTOR_HASH_REPAIR_SAMPLE_SIZE = 21 validators are malicious is roughly 1/1000. Assuming we send a separate sample every 5 seconds, that's once every hour.

On the other hand with a 52-48 split of validators with one version of the block vs another, the chance of >= 11 of the 21 sampled being from the 52% portion is about 57%, so we should be able to find a correct sample in a reasonable amount of time.